pyproject.toml
[project]
name = "crbn-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.127.0",
    "google-cloud-texttospeech>=2.33.0",
    "piper-tts>=1.3.0",
]
